add graphviz
